[BugFix] Validate template decoder section ranges
- Reject reversed and out-of-bounds CSS, style object, and Lepus chunk routes before creating sliced readers.
- Use size_t for decoder slice lengths and share section range validation across decoder paths.

issue: f-9723655
AutoSubmit: true

+++ core/runtime/lepus/binary_input_stream.h
@@ -117,7 +117,7 @@ struct InputBuffer {
 
 class ByteArrayInputStream : public InputStream {
  public:
-  ByteArrayInputStream(const uint8_t* data, int len) {
+  ByteArrayInputStream(const uint8_t* data, size_t len) {
     buf_.reset(new InputBuffer());
     buf_->data.assign(data, data + len);
     Initialize();
